move i18 components to common (#715)
