+++ packages/client/src/modules/common/components/native/ui-native-base/components/LanguagePicker.jsx
@@ -4,7 +4,7 @@ import { Text, Platform, View } from 'react-native';
 import { ActionSheet } from 'native-base';
 import SimplePicker from 'react-native-simple-picker';
 
-import { HeaderTitle } from '../../../../common/components/native';
+import { HeaderTitle } from '../..';
 
 export default class LanguagePicker extends React.Component {
   constructor(props) {
