+++ packages/client/src/modules/common/components/native/ui-native-base/components/index.jsx
@@ -6,3 +6,5 @@ export { default as InputItem } from './InputItem';
 export { default as SwipeAction } from './SwipeAction';
 export { default as ListItem } from './ListItem';
 export { default as MenuButton } from './MenuButton';
+export { default as LanguagePicker } from './LanguagePicker';
+export { default as Root } from './Root';
