+++ packages/client/src/modules/common/components/web/ui-antd/components/index.js
@@ -22,3 +22,4 @@ export { default as CardText } from './CardText';
 export { default as Table } from './Table';
 export { default as ListGroup } from './ListGroup';
 export { default as ListItem } from './ListItem';
+export { default as LanguagePicker } from './LanguagePicker';
