--- packages/client/src/modules/i18n/components/native/index.js
@@ -1,3 +1,0 @@
-// eslint-disable-next-line import/prefer-default-export
-export * from './ui-native-base';
-// export * from './ui-antd-mobile';
