--- packages/client/src/modules/i18n/components/native/ui-native-base/index.js
@@ -1,3 +1,0 @@
-// eslint-disable-next-line import/prefer-default-export
-export { default as LanguagePicker } from './LanguagePicker';
-export { default as Root } from './Root';
