--- packages/client/src/modules/i18n/components/web/index.js
@@ -1,3 +1,0 @@
-// eslint-disable-next-line import/prefer-default-export
-export { default as LanguagePicker } from './ui-bootstrap/LanguagePicker';
-// export { default as LanguagePicker } from './ui-antd/LanguagePicker';
